Standardize competitionId and shooterId to Long throughout
Replace Int with Long for competitionId and shooterId across navigation
routes, ViewModels, repositories, DAOs, and remote data sources. This
eliminates unnecessary .toInt() conversions and makes ID types consistent
with the data model. Also removes debug Log.d() calls from shooter result
screen and unused @POST import.

+++ app/src/main/java/se/kjellstrand/webshooter/data/results/local/ResultsDao.kt
@@ -8,11 +8,11 @@ import androidx.room.Query
 @Dao
 interface ResultsDao {
     @Query("SELECT * FROM results WHERE competitionsId = :competitionId")
-    suspend fun getByCompetition(competitionId: Int): List<ResultEntity>
+    suspend fun getByCompetition(competitionId: Long): List<ResultEntity>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertAll(results: List<ResultEntity>)
 
     @Query("DELETE FROM results WHERE competitionsId = :competitionId")
-    suspend fun deleteByCompetition(competitionId: Int)
+    suspend fun deleteByCompetition(competitionId: Long)
 }
